[RW-100] Fix date helper format test
